requirements.txt
charm-crypto
cryptography
pydantic
click
pytest
hypothesis
pytest-benchmark
msgpack
